MOSTLY fix collider desync for punch attacks it looks like this issue has something to do with animations immediately stopping or teleporting.
the most favorable solution, added in this commit, was to take every animation and extend the final keyframe by 1 frame. this is 100x better than it used to be but is still a problem; it messes up after like 5 swings. hopefully this is fixed by the animation overhaul in 0.14; because I will have to put it to rest for now.